feat: made script more robust

+++ manager.py
@@ -1,5 +1,6 @@
 #!/Users/aviaryan/miniconda3/bin/python
 import os
+from sys import exit
 from cryptlib import get_file_list, encode, update_file, get_key, decode
 
 FLAG = 'FLAG_FILE'
@@ -17,13 +18,22 @@ if __name__ == '__main__':
 		# new case
 		# or decrypted state in power fail
 		print('Not encrypted, encrypting....')
-		update_file(encode, get_file_list(), get_key())
+		key = get_key()
+		print('Re-enter key')
+		key2 = get_key()
+		if key != key2:
+			print('Keys don\'t match, exiting')
+			exit(1)
+		update_file(encode, get_file_list(), key2)
 		createFlagFile()
 	else:
 		# encrypted already
 		print('Encrypted, give key to unlock')
 		key = get_key()
-		update_file(decode, get_file_list(), key)
+		failStatus = update_file(decode, get_file_list(), key)
+		if failStatus:
+			print('You entered wrong key. FO')
+			exit(2)
 		os.remove(FLAG)
 		# decoded, wait to close
 		ans = ''
